Refactoring attribute imports (#40)
